Refactoring auth logic due to change of auth-service location
