+++ apps/endatix-hub/app/(main)/(auth)/login/page.tsx
@@ -2,7 +2,7 @@ import Image from "next/image";
 import LoginForm from "@/features/auth/use-cases/login/ui/login-form";
 import type { Metadata } from "next";
 import NewAccountLink from "@/features/auth/use-cases/login/ui/new-account-link";
-import { getSession, SessionData } from "@/lib/auth-service";
+import { getSession, SessionData } from "@/features/auth";
 import {
   Card,
   CardContent,
