+++ apps/endatix-hub/app/(main)/forms/[formId]/update-form-status.action.ts
@@ -1,6 +1,6 @@
 "use server";
 
-import { ensureAuthenticated } from "@/lib/auth-service";
+import { ensureAuthenticated } from "@/features/auth";
 import { updateForm } from "@/services/api";
 
 export async function updateFormStatusAction(
